Retoques Finales

+++ app/src/main/java/com/example/sketchtrain/dataclasses/Sets.kt
@@ -4,7 +4,7 @@ import java.io.Serializable
 
 data class Sets(
     var number: Number = 0,
-    var weight: Double = 0.0,
+    var weight: Int = 0,
     var reps: Int = 0
 ) : Serializable
